// src/types/navigation.type.ts
export interface ICategories {
  Desktops: string;
  LaptopsNotebooks: string;
  Components: string;
  Tablets: string;
  Software: string;
  PhonesPDAs: string;
  Cameras: string;
  MP3Players: string;
}

export interface IDesktop {
  PC: string;
  Mac: string;
  ShowAll: string;
}

export interface ILaptop {
  Macs: string;
  Windows: string;
  ShowAll: string;
}

export interface IComponent {
  Mice: string;
  Monitors: string;
  Printers: string;
  Scanners: string;
  WebCameras: string;
  ShowAll: string;
}

export interface IMP3Player {
  Test4: string;
  Test5: string;
  Test6: string;
  Test7: string;
  Test8: string;
  Test9: string;
  Test11: string;
  Test12: string;
  Test13: string;
  Test14: string;
  Test15: string;
  Test16: string;
  Test17: string;
  Test18: string;
  Test19: string;
  Test20: string;
  Test21: string;
  Test22: string;
  Test23: string;
  Test24: string;
  ShowAll: string;
}

export interface ILeftMenu extends ICategories {}